v3

+++ english_app/pages/viewWordsList.php
@@ -1,65 +1,74 @@
+<?php
+ob_start();
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>SNC-Table</title>
+    <link rel="stylesheet" href="../css/viewList.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 </head>
-<style>
-table {
-  font-family: arial, sans-serif;
-  border-collapse: collapse;
-  width: 100%;
+<script >
+function myFunction(id) {
+  let text = "Are you delete this word?";
+  if (confirm(text) == true) {
+    window.location.replace("../pages/action_php/deleteAction.php?index="+id);
+  }
+  
 }
-
-td, th {
-  border: 1px solid #dddddd;
-  text-align: left;
-  padding: 8px;
-}
-
-tr:nth-child(even) {
-  background-color: #dddddd;
-}
-</style>
+</script>
 <?php
-function printResult($maxData,$currentPageNum,$data,$totalNum,$index){
+function printResult($maxData,$currentPageNum,$data,$totalNum){
   for($i =  ($maxData*$currentPageNum) -5 ; $i < ($maxData)*$currentPageNum;$i++){
     if($i >= $totalNum){
       break;
+    }
+    if($data[$i]->isDelete == "Y"){
+      continue;
     }
     echo "<tr><td>".$data[$i]->word."</td>
     <td>".$data[$i]->chWord."</td>
     <td>".$data[$i]->type."</td>
     <td>
-    <a href='editWord.php?index=".$index."'>Edit</a>
-    <a href='../pages/action_php/deleteAction.php?index=".$index."'>Delete</a>
+    <a href='editWord.php?index=".$data[$i]->id."'> <button class='editBtn' role='button'>EDIT</button></a>
+    <button class='button-24' role='button' onclick='myFunction(".$data[$i]->id.")'>DELETE</button>
     </td>
     </tr>";  
-    $index++;
   }
 } 
 ?>
 <body>
-<!-- direct to add word page-->
-<a href='../pages/addWord.php'>Add Word</a>
+<div class="top"></div>
+<div class="btnPosition">
+  <a href="../main.html" >
+      <button class="btnQ" id="quitBtn"><</button>
+  </a>
 
-<!-- back to main page-->
-<a href='../main.html'>Back</a>
+  <a href='../pages/addWord.php'>
+      <button class="btnA" id="addBtn">+</button>
+  </a>
 
-<br><br><br>
+</div>
+
+<div class="searchBar">
 
 <form method="POST" action ="./viewWordsList.php?page=1&search=Y"> 
-<input type="text" name="searchWord"/> 
-<input type="submit" name="search">
+
+<a href="./viewWordsList.php">
+        <button id="reloadBtn"><i class="fa fa-refresh"></i></button>
+</a>
+<button type="submit" name="search" id="searchBtn"><i class="fa fa-search"></i></button>
+<input type="text" placeholder="Search" name="searchWord" id ="searchInput">
+
+
 
 
 </form>
 
-<!-- reset result-->
-<a href="./viewWordsList.php">
-        <button>Reload</button>
-</a>
+
+</div>
 
 
 <table>
@@ -73,9 +82,9 @@ function printResult($maxData,$currentPageNum,$data,$totalNum,$index){
       $data = file_get_contents('../pages/action_php/json/words.json');
       $data = json_decode($data);
       $totalNum = count($data);
+      
       $maxData = 5;
       $currentPageNum =$_GET['page'];
-      $index = 0;
       $searchWord = array();
       $word = "";
 
@@ -83,7 +92,12 @@ function printResult($maxData,$currentPageNum,$data,$totalNum,$index){
         if(isset($_GET['search'])){
           $wordNum =0;
           for($i = 0 ; $i < $totalNum ; $i++){
-            $string = $data[$i]->word;
+            if($data[$i]->isDelete == "N"){
+              $string = $data[$i]->word;
+            }else{
+              continue;
+            }
+            
             if(isset($_GET['word'])){
               $word = $_GET['word'];
             }else{
@@ -96,26 +110,38 @@ function printResult($maxData,$currentPageNum,$data,$totalNum,$index){
               continue;
             }
           }
-          printResult($maxData,$currentPageNum,$searchWord,count($searchWord),$index);
-          $pageNum = ceil($wordNum/5);
+        
         }else{
-          printResult($maxData,$currentPageNum,$data,$totalNum,$index);
-          $pageNum = ceil($totalNum/5);
+          $wordNum =0;
+          for($i = 0 ; $i < $totalNum ; $i++){
+            $string = $data[$i]->isDelete;
+            if (str_contains($string, "N")){
+              array_push($searchWord,$data[$i]);
+              $wordNum++;
+            }else{
+              continue;
+            }
+          }
+          
         }
+        printResult($maxData,$currentPageNum,$searchWord,count($searchWord));
+        $pageNum = ceil($wordNum/5);
       }else{
         header("location: ./viewWordsList.php?page=1");
       }
     ?>
 </table>
+<div class="numlist">
 <?php
   for($i = 0 ; $i<$pageNum ; $i++){
     if(isset($_GET['search'])){
-      echo "<a href='./viewWordsList.php?page=".($i+1)."&search=Y&word=".$word."'>".($i+1)."</a>";
+      echo "<a href='./viewWordsList.php?page=".($i+1)."&search=Y&word=".$word."' class='pageNumList'><button id='numBtn'>".($i+1)."</button></a>";
     }else{
-      echo "<a href='./viewWordsList.php?page=".($i+1)."'>".($i+1)."</a>";
+      echo "<a href='./viewWordsList.php?page=".($i+1)."' class='pageNumList'><button id='numBtn'>".($i+1)."</button></a>";
     }
       
   }
 ?>
+</div>
 </body>
 </html>
